Refactored file structure. edited nodemon start in package.json
