+++ petal-patch-server/package.json
@@ -10,8 +10,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "nodemon ./bin/www",
-    "start-server": "node ./bin/www"
+    "start": "nodemon src/bin/www"
   },
   "dependencies": {
     "cookie-parser": "~1.4.4",
